backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# LangChain
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.0
langchain-community==0.0.10
langchain-experimental==0.0.47

# Vector Databases
chromadb==0.4.18
faiss-cpu==1.7.4
pymilvus==2.3.4

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2

# Image Processing
Pillow==10.1.0
pytesseract==0.3.10

# Audio/Video Processing
moviepy==1.0.3
SpeechRecognition==3.10.0
pydub==0.25.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.1
mysql-connector-python==8.2.0

# AI Models
openai==1.6.1
anthropic==0.8.1

# Graph
networkx==3.2.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
